Fix backpropagation of conditionals when the type is later widened
Consider the following example:
```
f() = (1, 1)
f(s) = i == 1 ? (1, 2) : nothing
function foo()
    next = f()
    while next !== nothing
        next = f(Core.getfield(next, 2))
    end
end
```

This was causing unnecessary allocations, because inference failed
to inform codegen that the reference to `next` inside the loop was
guaranteed not to be `Nothing`. Ironically if `f()` could also have
returned nothing, inference was able to figure this out, thanks to
pass the result of the comparison was inferred as `Const(true)`,
and on the second pass as `Conditional(next, Tuple{Int, Int},
Const(nothing))`, the tmerge of which is `Bool`. Fix that by
implementing Jameson's suggestion in #25261, and instead infering
the first pass as `Conditional(next, Tuple{Int, Int}, Union{})`.
To make this work we also adjust the lattice relation to have
`Conditional(next, T, Union{}) ⊑ Const(true)`. The result is that
the result of the comparison does not get widened to `Bool` as
quickly, retaining the type information on `next` and allowing
codegen to emit efficient code.